Optionally skip git sync

+++ mongodb/scripts/bin/sync_repo_with_ws.sh
@@ -6,6 +6,23 @@
 RELATIVE_REPO_PATH="devel/mongo"
 REPO_PATH="$HOME/$RELATIVE_REPO_PATH"
 
+
+# Parse arguments
+SKIP_GIT=false
+
+while [[ $# -gt 0 ]]; do
+  case "$1" in
+    --skip-git)
+      SKIP_GIT=true
+      shift
+      ;;
+    *)
+      echo "Unknown argument: $1"
+      exit 1
+      ;;
+  esac
+done
+
 remoteExec() {
     cmd=$1;
     echo "Running on devel-ws: $cmd"
@@ -20,30 +37,32 @@ else
 fi
 goToRepo="cd ~/$RELATIVE_REPO_PATH/$REPO_DIR_NAME"
 
-# Get the branch and revision to sync them later on the remote device.
-pushd "$dir" > /dev/null
-branch=$(git rev-parse --abbrev-ref HEAD)
-revision=$(git rev-parse HEAD)
-popd > /dev/null
+if [[ "$SKIP_GIT" == false ]]; then
+    # Get the branch and revision to sync them later on the remote device.
+    pushd "$dir" > /dev/null
+    branch=$(git rev-parse --abbrev-ref HEAD)
+    revision=$(git rev-parse HEAD)
+    popd > /dev/null
 
-# Run `git checkout` to the specific branch on the remote device.
-remoteExec "$goToRepo; git checkout --quiet $branch;"
+    # Run `git checkout` to the specific branch on the remote device.
+    remoteExec "$goToRepo; git checkout --quiet $branch;"
 
-# If checkout fails, run `git fetch` on the remote device
-if [ $? -ne 0 ]; then
-    echo "git checkout failed."
-    remoteExec "$goToRepo; git fetch --quiet; git checkout --quiet $branch && git reset --hard $revision"
-fi
+    # If checkout fails, run `git fetch` on the remote device
+    if [ $? -ne 0 ]; then
+        echo "git checkout failed."
+        remoteExec "$goToRepo; git fetch --quiet; git checkout --quiet $branch && git reset --hard $revision"
+    fi
 
-if [ $? -ne 0 ]; then
-    echo ""
-    echo "FAILURE: Checkout or reset failed. Make sure the branch has been pushed to origin"
-    echo ""
-    echo "To push the current branch and set the remote as upstream, use"
-    echo ""
-    echo "        git push --set-upstream origin $branch"
-    echo ""
-    exit 1
+    if [ $? -ne 0 ]; then
+        echo ""
+        echo "FAILURE: Checkout or reset failed. Make sure the branch has been pushed to origin"
+        echo ""
+        echo "To push the current branch and set the remote as upstream, use"
+        echo ""
+        echo "        git push --set-upstream origin $branch"
+        echo ""
+        exit 1
+    fi
 fi
 
 # Sync all the files
@@ -55,22 +74,24 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-# Set the git repository of the remote device to the given revision.
-remoteExec "$goToRepo; git reset --soft $revision"
+if [[ "$SKIP_GIT" == false ]]; then
+    # Set the git repository of the remote device to the given revision.
+    remoteExec "$goToRepo; git reset --soft $revision"
 
-# Perform a fetch if `git reset` failed.
-if [ $? -ne 0 ]; then
-    echo "git reset failed."
-    remoteExec "$goToRepo; git fetch --quiet; git reset --soft $revision"
-fi
-if [ $? -ne 0 ]; then
-    echo ""
-    echo "FAILURE: Reset failed. Make sure the branch has been pushed to origin"
-    echo ""
-    echo "To push the current branch and set the remote as upstream, use"
-    echo ""
-    echo "        git push --set-upstream origin $branch"
-    echo ""
-    exit 1
+    # Perform a fetch if `git reset` failed.
+    if [ $? -ne 0 ]; then
+        echo "git reset failed."
+        remoteExec "$goToRepo; git fetch --quiet; git reset --soft $revision"
+    fi
+    if [ $? -ne 0 ]; then
+        echo ""
+        echo "FAILURE: Reset failed. Make sure the branch has been pushed to origin"
+        echo ""
+        echo "To push the current branch and set the remote as upstream, use"
+        echo ""
+        echo "        git push --set-upstream origin $branch"
+        echo ""
+        exit 1
+    fi
 fi
